Delete stale files
The .cproject and .project files are from an IDE that I don't use
anymore.

The Windows makefile doesn't function, and I'm not particularly inclined
to deal with building for Windows in the first place. The regular
Makefile should be useable by MinGW for user's on Windows. Any further
support will need to call on Microsoft's C++ tooling, and I don't have
any of that.

The `extras/` folder held an SVG that was once meant to be pushed into
the program for drawing. SVGs are more complex than I want for something
that's just tracing a vector of points.