Updates to Makefile
The file is *almost* completely new. In order, the changes are:

Drop the `CC` variable assignment. That variable is for selecting the C
compiler, not the C++ compiler. While it happens to work in many cases,
it's not guaranteed to do so, and may cause funny linkage issues
Somehow the actual compile target was written correctly to call on `CXX`

The `sdl2-config` tool was not being invoked by Make. I'll have to check
the manual to be sure, but I think it was trying to dereference the
variable named `sdl2-config --cflags --libs`. This is neither a valid
variable, nor does it contain any values at all.

`CXXFLAGS` and `LDFLAGS` are now having values appended to them, rather
than outright being replaced. Builders can pass in some settings, now.

A `clean:` target has been created. It's super simple and just deletes
the executable file. Now the README isn't lying to the reader! (oops)

An `install:` target has been created. It installs the program according
to the Filesystem Hierarchy Standard. There is a `DESTDIR` variable that
is entirely absent from the rest of the file. This is to be compliant
with the requirements of Debian packaging machinery. `dpkg-buildpackage`
will override the install path via the DESTDIR variable in order to
place the program in a folder that can be rolled into the package.

The description of using Make for iterative builds has been dropped. The
example targets are not particularly insightful, and I can find another
reference if I need it. This project could have been just a single file.
Iterative builds are hardly a concern.