Begin fixup of CircularBuffer
Drop the virtual destructor declaration. This will guarantee a buggy
container implementation, but the destructor was already empty. It was
already broken. The program currently can't deallocate the
CircularBuffer objects. I'll be changing the inner container to a vector
or the like so that it is easier for me, and correct by default.

The `::peek()` method has been set to return a reference to constant
Point2D. I don't need to copy the object out to give something a view.
But I *do* need to guarantee that the caller doesn't alter the object.

the `::push(...)` method has been changed to accept the Point2D by
value. I am guaranteed one copy because it needs to be put into the
list. This is not likely to incur an extra copy as the compiler can
almost certainly see through it. The copy into the function scope, and
again into the array will be optimized down to just one copy from the
caller into the array.