Make CircularBuffer bits more private
The inner workings of the container should be kept private, and made
visible through accessor methods. I must have done it this way because I
was trying to avoid extra function calls or something. Semantics are
more important for the programmer. Make the program correct, then fast.

Besides, more detail in the code will let the optimizer do better work.