Back CircularBuffer with std::vector, drop a const
I can drop the manually allocated array and exchange it with a vector of
Point2D classes. C++11 gives std::vector a `.data()` method, which gives
me a convenient way of accessing it's array. This is still at risk of
memory corruption, but I'm making progress.

I had to drop a `const` qualifier on the function. This trailing const
indicated that the function would not mutate the object. Unfortunately,
the pointer is used by the createDFT function in order to out-parameter
the chains. It is used *explicitly* to perform mutation. The const has
been dropped for now.