Internalize chain creation completely
The function is supposed to create the epicycle chains. Previously, it
was *initializing* them -- they already existed, just in a useless
state. Out-parameters are dumb, anyway, so get rid of it.

I'm not sure what I was imagining with the success status return value.
Something to do with whether or not the function encountered an error,
I suppose. The better answer would be some kind of `Option` or `Result`
type, or just retular old exceptions. There's no detection, and I need
the return slot for the chains, so the switch has been made.