Begin change to new `createDFT` interface
The signal is created as a vector of points, and it is used to create
the chain pair. The pair creation has been moved to be completely after
the default signal creation.

`createDFT` is too restrictive with the use of `const Point2D`s, so I'll
have to drop that. I might be able to figure out a way to create them
and push them into a vector, but I'd rather not at this time.