Don't specify user & group ownership in install
Non-root users can't assign files to be owned by the root user, so this
fails. Writing to '/usr/bin' is also restricted, so running as a non-root
user will *still* fail. Running it via `sudo` means the install already
runs in the context of the root user, so it's already the right permissions.

So don't specify the ownership at all.