Move the build-dir dependency to EXE target
The rebuilds were the result of the `build/` folder being "updated" when
it should not be. The mtime on the folder gets bumped to the current
system time for every single call to the make-build-dir target
`$(BUILD_DIR):`

The build dir will end up with an age of "just slightly older than the last
.o file to be made". Next time around, the first object file is
considered out of date and is re-built, "updating" the build folder
along the way. The second object file is checked, resulting in the same
decision -- thus everything is rebuilt every single time.