Generate & include header file automake rules
Do the same kind of source-to-build-artefact translation as the object
code, but for GCC-generated automake snippets. This lets Make know
whether or not a given object file needs to be rebuilt based on whether
or not the *headers* have changed, rather than only the C++ file directly
associated with the object file.