Begin Debian package creation
Check in all the things! I generally like to add code & files in smaller
chunks, but it doesn't really make sense in this case. Much of the files
are very nearly boilerplate copy, and the parts that aren't are pretty
self-evident -- like fixing the names or dependency.

Onward to a changelog, and then release.